requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
PySide2==5.13.1
requests==2.22.0
requests-toolbelt==0.9.1
shiboken2==5.13.1
urllib3==1.25.6
aiofiles==0.4.0
argon2-cffi==19.1.0
async-exit-stack==1.0.1
async-generator==1.10
atomicwrites==1.3.0
attrs==19.2.0
cffi==1.12.3
Click==7.0
colorama==0.4.1
dnspython==1.16.0
email-validator==1.0.4
fastapi==0.42.0
graphql-core==2.2.1
h11==0.8.1
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.2
peewee==3.11.2
pluggy==0.13.0
promise==2.2.1
py==1.8.0
pycparser==2.19
pydantic==0.32.2
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.2.1
pytest-env==0.6.2
python-multipart==0.0.5
PyYAML==5.1.2
Rx==1.6.1
six==1.12.0
starlette==0.12.9
uvicorn==0.9.0
wcwidth==0.1.7
websockets==8.0.2
zipp==0.6.0